Do not display "Returning...." when going back to the Main Menu
When the user decides to jump back to the Main Menu, the Settings Menu would throw a message that the user is 'Returning...'.  It might be possible to re-provide the message again - but with an actual meaning that could prove to be useful for debugging.

Further, the string is normally on the display for about a few milliseconds - the user will never see it.  But, the it is only possible if the host system is slowly processing the shellscript for whatever reason.