Revised the function name from MainSettingsMenu() to Main()
Revised the function name from MainSettingsMain() to Main(); the reason for this change is due to the direction of the class's direction being changed to hold ALL related settings function (including sub-classes) to separate the class functions that only focus on the main categories instead.  For example, there will now be a class to hold "Zip" function settings and another class to hold "7Zip" function settings.  Previously, or the main idea previously, was to have ZIP function settings and 7ZIP function settings all in one class.  If I were to continue with this direction, the Settings class will be a giant class and a bit convoluted and overwhelming to read.