Disallow the ability to access Zip, 7Zip, or Git if feature is not available
Disallow the user's ability to access certain features from the menu if they are not available within the current session.  Ideally, while these items are not shown to the user - when not available -, the items should also not be accessible either in the same premises.  Previously, it would had been possible for the user to select the item while the feature is not available, or shown in the menu list.  The user could still access the item if either by accident or already knew the key to the menu they wanted to access.