When displaying the project path provided, make the messages nicer to the user
When providing the project path to the user, allowing them to see the location, also provide a nice message indicating if the program is able to find the directory.  This will help the user to quickly identify that the program is able to see or not see the project's source files within the provided path.