Instead of displaying our own message to provide a new path, use the Enums instead
Instead of displaying our own message to "Please provide a new path:", we will use what is already available to us via the DrawWaitingForUserResponse() function.  We will use this function to display "Please Provide a New Path or "Cancel" to cancel".  This is more unified and we avoid minor issues later that could arise.