Instead of using the 'PAUSE' extCMD, use the GetUserEnterKey
Instead of having the user to interact with the PAUSE extCMD, we will instead of the user utilize the the GetUserEnterKey from the Logging class.  Thus, the message and feedback is recorded.  But, most importantly, we centralize on the features that currently exists within the application to provide a better feel and experience.