Revised the DrawMenuItem functionality, added descriptions to General Program Configurations
Revised the DrawMenuItem to allow a new parameter, "AddNewWhiteSpace".  Which will provide a new line to help separate the menu items such that it is easier to read - to the user.  Otherwise, it may seem really cluttered and hard to follow.

Further, added detailed description for the menu items in the General Program.  This will help the user know what is within that menu item before they access it.  Remember, we do not have simple universal icons, we must rely on text to convey what is going on within the settings.