Forgot to add some semi-colons, now my life is complete!
A crazy get gets on the bus and tells the bus driver "What do you do when I went pay the fair?".  The bus driver looked at the guy and said calmly "Stop the bus and let him off.".  But then the crazy guy reached down his pocket and brought forward to the bus driver's attention - a semi-colon and said "Wrong answer, because I cut the brakes with this semi-colon!" while smiling.

So that's my story for why I love semi-colons, excluding the story that I just modified from the movie Spy Hard.