It is now possible for the user to customize the Notifications
It is now possible for the user to specify the type of notifications that they receive from the program at runtime.  Thus, for example, if the user wants to be notified for all events - it is now possible for them to do such.  If, the user wishes to be notified only for errors or just successful operations, they may do so at their convenience.