Expunged the Legacy POSH Check
Expunged the POSH Legacy check.  Alas, this algorithm or check - simply does not work where it is.  Because this check occurs at the extreme bottom of the script file, the POSH has to read top-bottom with no skipping.  As such, anything that breaks syntax or if there exists commands that are not yet supported on earlier versions, then POSH will stop reading the script file before it even reaches this check.

To really compact this, we would have to do two things:
 - Change and control the flow as to how the script operate
 - Create a bootstrap script