Added a new enumerator, SystemInformationPowerShellEdition.
This new enumerator, SystemInformationPowerShellEdition, makes it possible to know - in general - what major version of PowerShell is presently running this program.  Because there exists, two versions: Core (latest iterations) and Desktop (known as Legacy in this program), we can try to depict what instructions are to be executed or how to perform a specific operation that will work with such POSH version.