Removed PowerShellEdition() and the enumerator SystemInformationPowerShellEdition
Because PowerShellEdition() checks if the client is using a legacy version or the latest iteration of POSH, the legacy Windows PowerShell Desktop will simply never be able to run this program in full.  Because powershell has to read - top to bottom, reading every statement within the program before it reaches the next command to execute - older versions will be unable to properly handle this program due to various reasons: OOP source design, newer CMDLets, dotNET Core functionality, and how the application is handling data.

Thus, there's no real point in having this functionality remain - if it is virtually not usable.