Revised $__changelogLimit from int to uint32 datatype
Revised the data type for the variable  $__changelogLimit.  While an integer is indeed sufficient, it still carries a negative range - which is unusable in our case.  Instead, using uint32_t, we can use all of the available bits when the left most is set to one.

Removed a negative check in the function SetChangelogLimit().