It is now possible for the user to quickly toggle the 'Show Hidden Menus"
It is now possible for the user to toggle the "Show Hidden Menus and Options".  I really did not want this setting to be in the user's face - nor waste as much space due to its existence.  Thus, the user can quickly adjust the setting without having to go within a submenu to enable it or disable it.