Main Menu now calls the Builder!
Even though the Builder right now does - functionally nothing at this moment, it is still possible to reach the builder at this time.  This will come in handy for when we are ready to test out the builder during the piece by piece stage!

We are at the final stage of the development of Version 1!  SO EXCITING!  :smiley: