PrerequisiteCheck() now supports logging!
If a resource cannot be found, it will now be captured in the logfile - so the user knows that something horrible had happened during the operation.  While a successful pass through the function will also include a new logged event, but only stating that all of the resources had been accounted for and are ready.