Cache the value of the current detected status of a particular item
Instead of having to research if the item can be used, for whatever reason, we should only check that status at maximum - once.  Instead, we are calling the same instruction multiple times - thus calling the logging functionality to get the same messages on-top of each other - causing repetition.  While in performance sakes, this does minimal impact, but for optimization sakes - this is needed.