Resolved the double file extension issue.
Corrected a problem were the datafile archive file would have a double file extension, such as "$projectName.pk3.pk3".  The reason for this issue was that the CreateArchive() function had already attached a file-extension to the name, but GenerateArchiveFileName() would also apply the file extension as well.  Because the CreateArchive(), in both 7Zip and Default Compress, already provides this functionality - there's no need to duplicate the process again.  Thus, the file extension capabilities from the GenerateArchiveFileName() had been removed.