Do now show that the local repo is being updated until after user pref. check
Instead of advertising to the user that we are updating the local repository with the master repository, we will first check if the user wants to have the local repository updated by checking the user's preferences.  If the user wants to have the repo updated, then we will advertise the step.  Otherwise, we will not advertise it.