Variables that are already strings, do not need to be adjusted to work as string.
I think I keep thinking of Batch, so adjusting them with quotes and marking them as a variable within the same instance seemed safe. While it still is safe in by itself, it is just a waste of time for the POSH environment and file size. The exception to this rule is multi-lined strings that contains one or more variables. I will leave them as-is, it's just not necessary to alter them presently.