[WIP] Added a new function, AccessDirectory().
Added a new directory named, AccessDirectory().  This function will provide a graphical shell showing the contents within the folder, thus helping the user to see the contents in a natural friendly environment.  This will be much more helpful to the user to see the data, know where it is, and to easily access the contents.

I am marking this as 'WIP' as I need to further test this function and further polish it, such as adding try\Catch statements and future proofing.