Retransformed NotifyUserOfLocations() to ShowProjectLocation()
Replaced the function NotifyUserOfLocations() with ShowProjectLocation().  Because of the previous commit 4356784c6fb902a130e4c1382146381eae02d98b it is no longer needed to show the user the report directories as it is done already for us - and better.  Instead, we will now show where the compiled project file is located, as well as invoking the GUI Shell to show the file specifically.