Append '-dev_' to mark that the build is a dev build.
In Bootless Star, I had previously marked projects by revision ID.  For example, ($projectName) + "-r" + ($revisionNumber).  Now with Git, it seems a bit odd to have "SHA" or the like.  Instead, we will just mark it as a 'dev' build - with the SHA1 attached at the tail.  Perhaps there's another - nicer - approach to this, but for now - this will do.