Added custom bells, if incase something went horribly wrong while playing the audio
If incase the audio cannot be played, I assume that an exception could be triggered.  When such event occurs, than the beeps and the bloops will be provided.

NOTE: Going by the Microsoft documentation, it does not appear to provide any exceptions, but I am adding them just incase.  If incase there no exceptions, then I will change the usage in a later time.