Resolved an issue where an error would still report as successful
Resolved an issue in __AccessWebSite() where it would falsely report a successful signal - when in fact it had failed to open the webpage.

Improved the footprint, by removing a a boolean variable (probably a byte gained back),  $statusSignal.  We will instead use the $operationStatus variable to hold our progress within the function.