[WIP] Created a new function; CreateProjectTemporaryDirectory()
This is tied to the ticket https://github.com/SibTiger/PowerShell-Compact-Archive-Tool/issues/9 .
In order to remove files from the project, we must do so in a temporary location that will not effect the user's original source.

Keep in mind that this functionality is not yet complete, more work is needed to complete this functionality.