CopyDirectory() will now allow special characters.
I had changed CopyDirectory()'s Copy-Item to use the standard '-Path' argument instead of '-LiteralPath'.  I had done this as a way to duplicate the data inside of the directory - yet also treat the inside of the directory - as the root.  Previously,  using LiteralPath, the root of the directory - would be a level-up, not down.