Removed the ability for the user to customize the Notifications within PSCAT
Personally, I am not a huge fan of limiting the functionality within the program - or any program for that matter.  However, I believe that this program should only be simplified such that it is much easier for the user.  While the functionality, prior, worked just fine under the hood, but as time moves on - this feature could confuse the users as well be difficult to maintain as evolution continues forward.

To further expand, a lot of programs today - will not really provide the user with option to mute some or all of the sounds, instead this can be done through the Operating System's sound mixer.  Because the Windows Operating System provides this ability for us, if the user decides that they do not want any audible notifications - just use the Windows Sound Mixer to turn off or adjust the volume.  As such, the sounds that the program already provided will remain active - but it is no longer adjustable to the user for simplicity sakes.