Removed the ability to customize the 7Zip's Multithreaded functionality
Removed the ability for the user to allow or disallow the Multithreaded functionality within 7Zip.  This was done to help keep maintenance cost low and to further simplify how the application interacts with the end-user.  By default, 7Zip already makes the determination if the multithreaded operations will be utilized within the host-machine for us already, originally I provide this option such that the user has the ultimate decision over this feature.  However, I believe that a huge user-base would prefer to leave this functionality as-is by default.  With that, there is no real need to have a feature available - when hardly anyone will use it, especially if it is supposed to be maintained during the application's evolution cycle.