Added a new function, Delay()
Added a new function, Delay().  This function will provide the ability to temporarily delay the program's runtime for a specific amount of time.  This is useful if we need to pause the program's progress such as waiting for another program to finish, a delay before the next loop iteration, or if we need the user to read a piece of information before proceeding on-ward.