Finished the splash screen design!  :balloon:
The Splash Screen is now finished and provides very basic information regarding the program.

Remember programs that people use to run that were dated back in the 80's and 90's (which some corporations still use), the would have their own cheesy start up screens - whelp here we are!  It's magnificent!