Created a new script, StartUpArguments.
This script will hold arguments (at this time are optional) in which can control how the application will perform and behave.  Presently, $programMode is the only argument available.  This parameter, while completely optional, determines if the software will start in either: normal mode, cleanup mode, or uninstall mode.  If the parameter is not provided, then the default operation is the normal mode.