Added soon to be functional Clean function
As of right now, I created a clean() function that - while empty - will soon provide the ability to delete files from the host's system.  Further, this function is now called within the main script.  Depending on the $programMode, the program may launch normally or run the clean functionality.  In addition, I had added the clean script to the compiler.