Restructured the cleanup code
I decided to make some changes  with the structure of the entire cleanup function.  Instead of separating the two cleanup instructions, which would contain some duplicated code - instead we will merge them together and only perform the special actions when needed.  For example, we will only thrash the user configuration file if we are performing an uninstallation - not when we are cleaning up.