We will now evaluate the Start-Process and encapsulated it in a Try\Catch
In order to obtain the Exit Code from the Start-Process, we will have to capture the Start-Process information and then evaluate the ExitCode from the stored object data.  Further, we will use a Try\Catch to assure that Start-Process is able to execute properly.  If the Start-Process was unable to execute the desired command, then the user will see information regarding the error.  This information could be useful for debugging purposes - or helpful to the user for support means.