Resolved an issue where the return from Call is an Object[]
Because of PowerShell's POP defaults to System.Object[] as a return type for any value, excluding booleans, we must further evaluate the returned object to obtain the exit code.

This took a bit to solve as I am used to PowerShell's OOP, where the functions return datatype are explicitly defined.