Added further checks; is POSH Core detected?
Before we can execute the PSCAT program, we have to also make sure that the POSH Core can be found before attempting to use it.  The detection, presently only covers $PATH.

Originally, I was going to let the try\catch when using the CMDlet, Start-Process, handle the grunt work for me - but that will not work as I hoped.  Instead of using the Try\Catch acting as our end-all issues, we will have to further refine our checks a bit more, instead.

This is still in a WIP phase, though the default behavior presently works as before.