Trimmed superfluous functionalities in Call()
I had learned that it is not really proper to use a Try\Catch conditional check when attempting to call an external command while using the CMDLet, Start-Process.  Because of this, there is no need to worry about exceptions nor the Try\Catch implementation.  We will have to tackle this differently.