Transformed program from POP to OOP
Today I learned that the 'return' statement's behavior is not like from C based languages.  Because of this, and much headache, I had decided to spend less than thirty minutes to transform this entire shellscript to Object-Oriented Program.  I knew that using OOP I could get the return statements to work as I needed, which is what led me to quickly convert this script.

:tired_face: