Start-Process is, once again, encapsulated with try\catch
To make sure that the Start-Process CMDlet executes thoroughly, I had encapsulated it with the Try\Catch conditions.  This will make sure that if there was an issue with the Start-Process, then the user will now what might had happen and the information that was used to invoke POSH.

The intention of using the try\catch is to make sure that POSHCore was started successfully, anything past that - is not detectable to us using Try\Catch, other than inspecting the return code and inspecting the value.