Further refined the error detecting during and after Start-Process
If Start-Process could not launch POSH correctly, then we will have an error message that states exactly that.  Further, if PSCAT  could not launch correctly or the arguments given to POSH were not correct, we will also receive an error message for those scenarios.

This should help let the user know what is happening within the operations; did their request go through successfully or did an error reach?  Feedback is the key here