Added the ability to warn the user about deleting Program FIles
Added the ability for the Uninstaller to let the user know that, they - the user, must perform the manual deletion operation.  This cannot be done programmatically at this present moment.

The uninstall approach can be done for the user, but it is going to require so much more work and effort where it may not be beneficial for the cost of developing that feature.  Further, if users prefer this portable installation method - they are most likely Power Users and are much more comfortable with handling special actions - as long as it makes sense and are easy.

Perhaps in the future I can come back to this and add in this feature.