Reorganized the project's assets
Because this project is continuing to evolve, evolution cycle, we need to make sure that our workspace is evolving with it - in a clean and organized way.  With that, we need to make sure that we can easily navigate and minimize confusion as I continue to work within this environment.

I guess this was long overdue, but well worth it :)