Continuation of renaming Notifications --> NotificationAudible
NOTE: Originally, I was thinking that renaming the file via GitHub - would maintain the previous history.  However, that was not the case.  Regardless, the show must go on!

As a continuation of commit 582f13a4edfce800585d80cef4940d7c5b94ee16 , this provides the necessary changes to shift from a generalized notification object, to an audible notification object.