Changed the data type of $debugInfo to be generic
Because previously I was using $debugInfo as being explicitly tied to ApplicationInfo, it was causing conflicts when it came to different data types, rightfully so.  Because I am using DetectCommand() for various Command Types, it is best to allow the $debugInfo to flow freely with different data types that Get-Command provides us.