Changed string array to an arraylist.
Changed the String array data type to an Collections.ArrayList type.  The reason for this change is due to the multiselect being an infinite number.  Because the standard array requires a fixed size, we cannot predict the amount of files that a user will select.  Instead, we want to assure that whatever file(s) the user selects - we can easily support it.