Added a new script, Launcher.
This new script is in response to the ticket https://github.com/SibTiger/PowerShell-Compact-Archive-Tool/issues/34 .  In order to work around this complication - and as painless as possible, the startup sequence is going to have an extra step.  To elaborate, PSCAT has to be launched by a Launcher (this script that is added), such that the environment can be prepared for the PSCAT application.  Once the environment had been initialized, PSCAT can then run normally.  By doing this approach, I can assure that the features are available to the user at run-time.