Resolved a nasty issue where PSCATs dependencies were not loaded
The main fix was changing the **Start-Process** CMDLet to **Invoke-Expression**.  What this essentially means, PSCAT was started in a new environment - where the libraries were not included.  But using Invoke-Expression, I am able to use the exact same environment - where PSCAT is loaded on-top of the launcher's instance.

With this change, we do not need to worry about finding nor calling *PowerShell Core*.  Thus, the extras are no longer necessary and had been expunged.