Provide a message box to the user after the clean\uninstall operation
A message box will now be provided showing the final results of the operation.  Thus, if the operation was successful, then the message box will display a successful message.  Otherwise, if a failure had been reached - then that message box will contain an error message along with the directories that could not be removed.