Harden BrowseForTargetFile()
This function will no longer blindly accept any valid path as a legitimate source.  The problem with the previous implementation, it is possible for a directory to be marked as a file.  This was possible because the path - was valid.  Now to help minimize potential user errors, we will only allow actual 'files' to be valid.  Directories that were mistakenly provided will no longer be acceptable.