Created a new script file that will soon be the centralized hub
This script is soon to over take the Inno Configuration script file.  The idea is to have multiple scripts that focus on their own regions, which makes it easier to maintain and evolutionize as time goes on.  Otherwise, if we keep using just one file to hold every configuration possible - it will be extremely hard to upkeep.