[WIP] Installer will now inspect if POSHCore is installed
The installer will now check if the PowerShell Core application had already been installed on the hosts system, if not - the installer will open the user's preferred webbrowser to obtain the latest official POSHCore installation package.

NOTE: This is very WIP and breaks the main installation.  This is more of a proof of concept; assuring we have all the tools in the toolbox to make this feature possible and working.