[WIP] Overhauled the Detection Algorithm
We will use recursion to make this appear a bit necessary and slightly easier to work with in regards to splitting the code in to two parts.  The first part is finding all of the possible subkeys, while the other inspects what we require.

This Algo is not yet finished and still has some leaks to it.