Revised how the algorithm will operate on different CPU Wordsize
Changed how the algorithm will work solely dependent on the Operating System's support on the CPU wordsize.  This helps to further simplify the logic and instructions.

Further, this resolves an issue where the scanned results - despite being positive - is treated as negative.  Thus, regardless of the results, the scan would always report that the dependency was not found.  This seems to only effect if the 32Bit application was only installed within a 64Bit Operating System Environment.