Changed my mind, not supporting a Custom Uninstall
Moving forward without this functionality.  I am not too sure how to achieve exactly what I want due to a limitation of knowledge, documentation, and useful examples.  Further, I am spending too much time on this - that the overall benefit is simply not worth it.

Moreover, presently, all directories and files that are associated with the PowerShell Compact-Archive Tool will be deleted upon uninstallation, only exception to this rule is the Compiled Builds.  The Compiled Builds will NOT be deleted upon uninstallation, the user must do so manually.