Do not ask the user where to install PSCAT
We will no longer prompt the user as to where the contents of PSCAT should be installed.  The reason behind this - is due to PowerShell's default limitations.  When the user clicks on a PS1 script, by default, Notepad.exe opens with the shellscript in the editor buffer.  Instead, we will have to create a front-end way to run the PSCAT application without needing to alter the user's settings.  This - honestly - is more of a question as to why Microsoft decided that this was a good idea, when one can clearly cause havoc on an individual's computer by means of a Batch shellscript or even an Ol' fashion .exe, which - NONE of what I just mentioned are blocked from standard execution.  What I mean from all of this, I don't understand why PowerShell has this strange limitation when it is simply not needed.  Further, it makes legitimate programs - like mine, harder for the user to utilize in an OOBE fashion -- should only need to click and go, no need to make the user's experience harder due to shotty reasons.

So, with that said, again - I am going to make shellscript to help launch the program.  To help simplify my work, I am going to make some Batch shellscripts to run the program using POSHCore.