Deleted URL Shortcuts as Inno provides a nicer way already
Originally, I did not know that Inno Setup can provide us with the URL Shortcuts in the Start Menu when specified in the scripts.  Because of this, I believe that having the links in the Start Menu - thanks to Inno - would be much better than having the links also in the Installation Path.  If keeping the shortcuts in the Installation Path, it would be wasting space - despite being extremely small of 5 or 6 KBs, it would have absolutely no meaningful value.