Uninstaller will no longer be accessible in the Start Menu
The Uninstaller is still accessible in the Control Panel's 'Programs and Features' (Windows 7 and later) and the 'Apps' in the Metro Control Panel (Windows 8 and later).

The reason for this change is due to potential misclicks.  Further, it seems much more appropriate to have the uninstaller located with the rest of the uninstallers as well.  No need to clutter the space that the user visits frequently.