Created an empty file, Bootstrap Loader batch shellscript.
This will be used, once populated, to start up the PSCAT program without the user needing to configure their system due to PowerShell's strange and inconsistent default settings.

The reason it is empty for now is because I could not create this file on my mobile due to Git's behavior with directories.