Removed Program Modes directory
Removed the Program Modes directory, including: Clean.ps1 and Uninstall.ps1.  The reason for this change is simply due to the Batch scripts - through the Windows Installer.  As such, the user can simply use the Batch scripts to easily launch the PSCAT program, regardless if POSH scripts are allowed to be executed within the user's environment -- bypassing the Set-ExecutionPolicy.