Removed ShowHiddenMenu
Removing the ShowHiddenMenu availability from the program.  I want to phase this out as it is a half-backed idea - or not a proper solution that should be using within the program.  Instead, we should smart as to what should be hidden and visible.  Only hide what the user should not see, but make the essentials available regardless.

NOTE: You may need to re-create your configuration file as __showHiddenMenu variable had been expunged from the program entirely.