Revised the Internal Zip Settings Menu
Revised the menu such that the user is now able to control all Internal Zip settings centralized within the menu respectively.  Further, the menu will now alert the user if the Internal Zip functionality is not being used.  This will allow the user to switch from one compression software to another.  When the tool is not presently used, all associated options will be hidden until the tool is active.

Added an extra layer of error handling if incase the Internal Zip is not present within the session.  This is only intended to alert the user that their PowerShell Core installation could be corrupted or missing required dependencies.