Revised the 7Zip Settings Menu
Revised the menu such that the user is now able to control all 7Zip settings centralized within the menu respectively. Further, the menu will now alert the user if the 7Zip functionality is not being used. This will allow the user to switch from one compression software to another. When the tool is not presently used, all associated options will be hidden until the tool is active.