Removed Compression Tool Selector from the General Program Settings
Removed the Compression Tool selection from the General Program Configuration menu.  Instead, now the user can be able to toggle the appropriate compression software directly from the Internal Zip and 7Zip menus respectively.

In doing so, this should be a much more intuitive approach to setup the compression software to the user's liking, and less of a Mouse Hunt game to figure out where to go in order to enable the desired compression software.  Now, the user can be able to enable the desired compression software directly from the appropriate compression software menus.