Removed Git Toggle Setting from the General Program Settings
Removed the Git toggle setting from the General Program Configuration menu. Instead, the user will soon be able to toggle the setting directly from the Git Configuration Menu.

In doing so, this should be a much more intuitive approach to enable or disable the Git utilization directly within the Git Configuration menu.  With that, it is Hide and Seek game as to where to toggle the Git features.  Soon, all of the Git functionality settings will all be stored in just one Menu, instead of some settings being relocated in other menus.