Revised the Git Settings Menu
Revised the menu such that the user is now able to control all Git-SCM settings centralized within the menu respectively. Further, the user can be able to enable or disable PSCATs use of Git-SCM - entirely.  Thus giving the end-user full control if the Git software will be used or not.