Removed User Option, Use Windows Explorer
By default, hardcoded, PSCAT will always use Windows Explorer for pure simplicity to the users.  The only time when Windows Explorer will not be used - depends on the Operating System.  If using any Operating System other than Windows, due to the nature of PowerShell Core and dotNET Core, the GUI elements will not be provided.