Vastly revamped how Version Control is managed
It is now possible for PSCAT to support other SCMs such as SVN, Surround, even Mercurial, but also any other version control tool that has Command Line capabilities.  With that said and done, one will have to design the Interface such that the functionalities can be actually supported within the program.  If one is willing to go that far, it will provide a way for PSCAT to openly support it.

Further, you may need to re-create your configuration file as __useGitFeature variable had been entirely removed and replaced with an enumerator variable __versionControlTool, enumerator datatype UserPreferencesVersionControlTool.