Always Update User Configuration when Settings Change
Always update the user's configuration file when they change a setting within the General Program Menu's.

With this change, this assures that the user's configuration file is always up to date after making the appropriate changes within the desired menus. Otherwise, previously, it could be possible that the user's configuration file could never be updated if the following had occurred:
- Program randomly crashes at some location other an terminating correctly. Normally, saving the user's configuration file only happens when the program is closed properly.
- User presses the 'X' button at the terminal window's controls.
- Terminal instance was abruptly terminated.