Relocated _LOGGINGLOCKKEY_ from Global to an Object
Relocated the _LOGGINGLOCKKEY_ variable from a global variable to a member variable within the Logging object.  With this, it was appropriate to rename it from _LOGGINGLOCKKEY_ to LoggingLockKey.

This also corrects a warning that the _LOGGINGLOCKKEY_ was assigned but never used.  Despite it was used previously, it seemed to trip a false-warning?  Nonetheless, it is corrected with this change.