Relocated Update PSCAT and About PSCAT to Preferences
Relocated the PSCAT Update and About options from the Main Menu to the Main Settings instead.  The reason for this - is to reduce clutter from the Main Menu, as not many user's will want to access the 'About' or 'Update' options all the time or frequently.  As such, this will just move those options out of the way and only have the user's focus on what is critically important to them.