Check if Project Sites are Available
Check if the project's sites are available to access.  If the project's website URL is not valid OR is not accessible on the host's system, then the options will be hidden from the Main Menu.

With this change, the CheckSiteAvailablity() function as well as the __CheckSiteAvailability_SiteResponse() function signature to incorporate the ability to suppress error messages.  With this ability, we can suppress error messages being provided to the user when inspecting the project's website(s).  As a benefit of this feature, if the host system has spotty internet, the options can become available when an internet connection is active.  Thus, the user does not need to restart the program to access the desired website(s).

Resolved a HUGE syntax error in LINE333 where the [WebsiteResources]::__CheckSiteAvailability_SiteProvided() function call incorporated an argument - but I the conditional check was not valid.  Thus, it meant really nothing.  HOWEVER, further, the entire conditional check was still passable as the function would return true if the SiteProvided was valid.