Revised the __DrawMainInstructions() documentation
Revised the documentation header for the function name, __DrawMainInstructions().  With this change, I provided more depth for the documentation section that outlines the purpose of the function.  In addition, I had also provided an Input section that provides a reason for the function arguments.  This should be enough to let the programmer know as to what is happening within the function.