Added WaitForProcess()
This function will halt the program until the specified Process ID, provided by the Operating System, had been released \ terminated.  If the application - in which the Process ID is belonging to, is frozen or cannot be released back to the Operating System, this function will not allow the program to run normally.  Assure using this function can return back to normal operation, otherwise this program can show effects of it being frozen - despite it is  waiting for the PID to be released.