Added WaitForFileExplorer()
This function will wait for a specific File Explorer window to be closed by the user or host, which will then allow this function to return the program back to normal operation.  If the directory is never closed by the user\host, then this function will prevent the program from running.  Otherwise, once the window had been closed - then this function will return normally.