[WIP] Overhaul ProjectInformation
NOTE: From here on out, PSCAT is not going to work correctly for awhile!!  Bare with me as I make these changes.

Overhauled the ProjectInformation to work as a single instance to objects as needed.  This change is going to be necessary for single instance of PSCAT using multiple projects, and how Embed Installer will function.