Warn the user to close folder window to continue
Alert the user to close the temporary directory folder instance in order to continue with the procedure.  Otherwise, the user will be confused by seeing "Press the Enter Key to continue. . .", yet the program is not waiting for keyboard input from the user - but to close the folder UI instead.