Revised verification shown results
The verbose information will now be only logged into the program's logfile, no longer visible to the user - allowing minimalistic feel.

Only show the user that the verification operation is on-going, but the detailed actions will remain internalized in the logfile.  If an error occurs, such as corrupted zip files, that will remain visible to the user as previously.  This is important - as it highlights to the user of a bad zip file.