Added EMbedInstallerFile constructor for legacy means
This constructor is merely for legacy means for the time being and will most likely be removed later on during the developement phase.  This is only added as the current process for the EmbedInstaller requires using a constructor that uses two arguments.  The later iterations of the constructor will require more information, but to keep development progressing forward - this temporary means will remain for the time being.