[BROKEN] Adding File Browser to Select Files
NOTE: This state  is currently broken; more work is needed.

Changing the way the user selects what files are to be uploaded into PSCAT's Project Manager to install within the environment.  Previously, the user would put project files into a temporary directory - and that would be the way to progress with what files are to be installed.  Instead, now - slowly, we will use the default Windows File Browser instead.  Using this functionality will be much more natural to the user and avoids possible confusion that can lead to files being accidentally expunged from the Windows environment.

Discussion of this change is here:
https://github.com/SibTiger/PowerShell-Compact-Archive-Tool/issues/68